requirements.txt
# Zero Side Effect backend dependencies

fastapi
uvicorn[standard]

sqlalchemy
pymysql

pydantic
pydantic-settings

python-multipart

numpy
Pillow
tensorflow
